test: separate project routing parser tests
