+++ cli/python/base_github_projects/tests/test_engine.py
@@ -110,70 +110,6 @@ def test_parse_project_configure_accepts_copy_fields_from_project() -> None:
     assert args.copy_fields_from_project == "Base Roadmap"
 
 
-def test_parse_project_issue_fields_defaults_to_repository_named_project() -> None:
-    args = engine.parse_args(
-        (
-            "project",
-            "issue",
-            "set-fields",
-            "1862",
-            "--repo",
-            "basefoundry/base-cli",
-            "--status",
-            "In Progress",
-        )
-    )
-
-    assert args.project_title == "base-cli"
-    assert args.repo == "basefoundry/base-cli"
-    assert args.allow_cross_project is False
-
-
-def test_parse_project_issue_fields_rejects_mismatched_project() -> None:
-    with pytest.raises(engine.ProjectUsageError) as excinfo:
-        engine.parse_args(
-            (
-                "project",
-                "issue",
-                "set-fields",
-                "1862",
-                "--project",
-                "base",
-                "--repo",
-                "basefoundry/base-cli",
-                "--status",
-                "In Progress",
-            )
-        )
-
-    assert str(excinfo.value) == (
-        "Project 'base' does not match repository 'basefoundry/base-cli' "
-        "(expected 'base-cli'). Use '--project base-cli' or pass "
-        "'--allow-cross-project' for an intentional cross-project update."
-    )
-
-
-def test_parse_project_issue_fields_allows_explicit_cross_project_update() -> None:
-    args = engine.parse_args(
-        (
-            "project",
-            "issue",
-            "set-fields",
-            "1862",
-            "--project",
-            "base",
-            "--repo",
-            "basefoundry/base-cli",
-            "--allow-cross-project",
-            "--status",
-            "In Progress",
-        )
-    )
-
-    assert args.project_title == "base"
-    assert args.allow_cross_project is True
-
-
 def test_project_auth_exit_code_is_named() -> None:
     assert engine.PROJECT_AUTH_EXIT_CODE == 3
     assert "return 3" not in inspect.getsource(engine.run)
